Fixed the bug causing segfaults in database.pyx
RecomputeFilters was assuming that the order of a sub-filter was always
the same as that of it's parent. When this assumption proved false, objects
were added to the database multiple times, corrupting objectLocs and resulting
in a segfault down the line when objectLocs was used.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@298 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 85ed001cc1c8119991fefc035f069bc03fb86559