Linux port fixes.
* Fixed crash/lockup when xineGetPosition is called bfore xineAttach.  There
was a race condition that would sometimes trigger this when network/processor
load was high.

* Couple of fixes for the play/pause button.  It should represent the current
play/pause state better now.

* Changed the behavious of the time slider and seeking in general.  It used to 
continously change the current time in the playback, while the video was
playing.  This caused a bunch of pops in the audio for me, and it probably
would be much worse when the media didn't allow fast seeking.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1055 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: a09994397835b9fb98edc93b5159caadb8524487