* Fixed dumb typo in restoring downloads.  I think this caused movies to not be saved once they were downloaded.
* Switched the way the download daemon works.  Now the main democracy process
opens a port, then tells it to the child and the child connects.  This
simplifies things, makes it faster and I think it might be more robust.  Now
that the roles are switched, I changed the names from "Server" and "Client",
now it's "Controller" and "Downloader"

* The Downloader Daemon now locks its PID file.  This should prevent more than
1 from ever starting.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1298 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 5bc2f12708ed2c4ee763b58ba221e03c48fabbf5