Added better error handling the the downloader daemon:
* Added a command for the downloader to send to the main process telling it
to pop up a error dialog.
* If the downloader gets an error communicating to the controller it quits
* If the controller gets an error communicating to the downloader it kills
the current downloader and restarts a new one
* If an error occurs in one of the download threads in the downloader daemon 
we should get an crash report for it.
* Replaced some blanket catch statements with ones that specify socket.error.
We 

Fixed a bug in item where if a feed had 2 identical enclosures the it could
hang.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1344 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f6dae9442abd23e56c3d961ad5f45f1432739910