Changed @gtkMethod to @gtkAsyncMethod
Reimplemented @gtkAsyncMethod using a very similar concept, but using a Queue instead of a [] and having each object from the queue in a separate callback function so that we don't block the UI as much.

Implemented @gtkSyncMethod which allows for a return code, but blocks until the idle call has finished.

Switched all users of @gtkMethod to @gtkAsyncMethod.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1407 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 89cf1b1c6beb906e5139a70ce44e9fd8efbd1298