Cleaned up the way downloader classes, especially how they get stored in the database
* Merged Downloader and RemoteDownloader class.  RemoteDownloader was the only
subclass of Downloader so it was sort of redundant.  If we wanted to add a
second subclass of Downloader I think we need to do a good amount of redesign
anyway.

* RemoteDownloader stores many less attributes.  Almost everything is stored
in the status attribute, which is the last value that we got from a
StatusUpdate message from the downloader daemon.

* RemoteDownloader now restores it self from the database.  Before we created
a new object on the daemon side every time we started up.

* Changed the way upgrades worked slightly.  Since we will always modify the
list of SavableObjects in place, there's no need to return it at the end.

* Fixed some minor bugs in the database tests.

* Added some extra validation checks for if an attribute is missing when we
save/restore the database.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1417 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: bae517e8f4a1033882ddaa337cf05a01c07dc3e5