Fix for #983.  We now have 2 parameters, DOWNLOADS_TARGET and TORRENT_DOWNLOADS_TARGET.  DOWNLOADS_TARGET controls the maximum number of auto downloads that we create, TORRENT_DOWNLOADS_TARGET controls the maximum number of auto downloads for torrent files.  DOWNLOADS_TARGET is 4, while TORRENT_DOWNLOADS_TARGET is 3.  This prevents slow downloading torrents from blocking http downloads.
The new strategy for auto downloading is to to 2 passes.  On the first pass,
we refuse to create more than 1 bit torrent auto-download.  On the second pass
we download whatever.  Both passes are still limited by DOWNLOADS_TARGET.
This makes us prefer http downloads, but if there aren't any we will download
torrents.

I cleaned up the autodler.py module a little while I was working with it.  I
made the DownloaderFactor try doing a HEAD request before a GET request.  If
it resorts to the GET request it closes the file-handle.  I think this was
causing us to download things twice.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1443 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 801f3226b5f485bff2140c974f1223f22ff1038f