Added support for videeos/torrents opened with democracy to change the tab. Torrents bring you to the download tab, videos bring you to My collection.
Refactored the controller a little to support this.

Moved the logic for selecting the tab out of the controller and into
singleclick.

Fixed a bug in commandlinehander.py that was preventing democracy from
launching.

Made the remote downloaded send a couple extra status updates.  The most
important one is immediately after we know the info hash, but before we have
started the torrent.  This helps to prevent starting duplicate torrents when
the user opens one twice.

Made it so we check file associtions that already handled by democracy in the
installer.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1529 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 48f8fb05b8cd3c98157fc5f015e1db215e63f1e7