Give the saveDatabase timeout a name.
If there's an IOError during database save, give a more useful message, especially since it's not a bug in the program, but a problem the user has to fix.

Do the database backup earlier.  It's the same if the database save fails, but this way, if the database save succeeds, there's a backup of the previous state.  Maybe we should only do the backup during the first save of a session, so that if it doesn't load the next time, there's still a useful backup?


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1625 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: ba6bd6be9bc57bd97a4864b92632c1e62a195574