Lots of updates to the http client.
I think we can replace calls to download_utils.grabURLAsync with
httpclient.grabURL.  The only difference I know if, is httpclient.grabURL
takes a callback and an errback, instead of returning None in the callback on
error.

Added a ConnectionHandler class that we may be able to use for the download
daemon communication.

Added coverage analysis for linux tests.

Added util.trapCall() which is an easy way of catching error in callbacks.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1635 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f936ba8798d714b48811ac2f33a795a3ae88d30f