Always have app.py migrate downloaders.  Let the lower levels decide whether or not to do the migration.  This avoids some potential race conditions.
If we have a dlid for a remote downloader, send the migrate signal to
the dl_daemon and assume that it's going to handle it.  We basically
have a dlid if there's a remote object and it's in either the
downloading or uploading state.  The other side will only migrate if
it's in the uploading state or finished state, but we send it always,
in case it goes from downloading to finished while we're sending this
command.  If we don't have a dlid, try to do the move.  We don't wait
for a response.

In the download daemon, we're no longer going to send a response since
the app isn't waiting for one.  Then, only move the download if it's
in the finished or uploading state.  We don't want to move it to its
final location if it's not done yet.  This used to be handled higher
in the stack, but if we handle it here, we don't have race conditions.

There's still the race condition that we never update the config in
the dl_daemon.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1662 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 7c39fef06c2443805e56e325177d62c6a9c43655