Fixed a bug in the HTTP code, if we closed the connection before the socket opened we would get exceptions.  Added test cases for this.
Separated the expired items check from the auto downloader check.  Checking
every item takes a (relatively) long time, and it's not very urgent, so I made
that timeout every 5 minutes.

Made the autodownloader asynrchronous.  This was pretty weird because we need
to to a HEAD request to check if an item is a torrent or not.  I think the
code seems to be working though.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1810 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 76eee95700ae4ad6faaef04a6f0a7822d7f3c605