Added save failed dialog for windows.
Added a new Dialog subclass, MessageBoxDialog.  Windows uses it for it's save
failed, but it's not used in portable anywhere.  I'm hoping to remove the
saveFailed method and just use this dialog eventually, but I don't want to
force other frontends to implement a new dialog this close to the release.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1835 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 679e65272cee7d9f31c22f9297bbf4fb8de01a67