Fixed the HTTP auth code with canceling requests.  HTTPClient objects now have a cancel method, if we have a request it'll cancel it.  If we're waiting to get HTTP Auth info, then we'll remember that we were canceled and do nothing when it comes back.
Fixed a bug in the cancel code, when a request became free we would set it's
requestId to 0.  This screws up pipeline-able connections, because they are
free as soon as they get the headers.

Fixed a couple of the http tests.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1936 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f38acdfc52753931234c0eaf29bd5ab8bb9f9078