Updated the remove channels dialog, hopefully the final time.
Fixed canceling requests and pipelining.  I added a callback in HTTPConnection
for when it starts a request (more precisely when it starts reading in a
response for a request).  HTTPClient registers for the callback, and keeps
track of which connection it's using.  If it gets a cancel() message, it
closes its current connection.  If the connection hasn't started yet, then it
closes it when it gets the callback.  I think this should fix the problem.

Added tests for the bugs.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1948 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 732b28a2342715b91585b1ce669db629a64fbf2f