Added svn:ignore to a couple directories.
Fixed several BitTorrent problems.  I finally gave in and fixed the bug where
bit tornado calls getmtime for a file that doesn't exist.  I still have no clue
how it's happening, but wendell has gotten it a bunch of times and I just got it
on the powerbook.  Made it so we wake the bit torrent event loop up when we
first add the get_status timeout before it was taking a minute or so before it
got around to it.  Added code to check for bit torrent errors in get_status.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/branches/Democracy-Player-0.8.5@2341 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: c0d6794fce0e5d947e4197780d7c9f71aa78bbbf