Reworked addFeed and addChannel to accept None as the url and put up their own dialog in that case.  Switched gtk frontend to call these functions with None.
This uses prefillCallback for if the user types an invalid URL.  This callback wasn't implemented on gtk, so it may not be implemented on the other OSes.

None of these changes are necessary.  The portable code will continue to work with the old platform code, but the code will all be simpler if the platform code uses this code.  And if prefill isn't implemented, they'll just get an empty box instead of the invalid url they already typed.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@2594 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 54e098e8de7f890f18014be12aab1eda4ddfb082