Adding support for playlist reordering.
Fixed a potential bugs with the database code.  I noticed it when I was
trying to track down some segfaults when we reordered playlists.  If the
cursor is on the first object in the list and we remove that object, the
cursor should point to None after.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@2629 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 71ae3c3a0dd7a54813b05b92ee044656ce3a05e4