Added templates for mulitple playlist/folders selected.  Made it so when a playlist is created we select it.  Fixed crash when adding a channel guide on gtk when the user gives a bad URL.
The big change:  Removed the currentTab attribute.  Since multiple tabs can be
selected, this should be a list.  I noticed that we accessed currentTab from
the GUI thread in several places (onURLLoad, validating menus).  This may have
been okay when there was a single tab, but it will definitely cause trouble
with multiple selected tabs.

To fix this, I added a method to MainFrame called onSelectedTabChange().  When
we switch tabs, we call this from the portable code and pass it the type of
tab that's selected and if multiple tabs are selected or not.  This should be
enough info to validate the menu items.

We also pass it the url of the selected channel guide if a channel guide is
selected.  Then the portable code sets the guideURL attribute in on the
Controller.  Finally, in onURLLoad we check this attribute to decide if a URL
is coming from the channel guide.  This seems like a fairly ugly solution to
me, but I can't think of a cleaner alternative.  If someone can, please do it.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@2661 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 5fb4b87e528a418eea621757f195a119d69860e0