First pass at tab reordering.  I created an object called TabOrder that stores the order of the tabs.
Slight change to DND code to make handleDrop get both the item being dropped
on and the type of drag we're doing.  This is for playlists, and folders that
have 2 kinds of drops that can happen on them..  If you drag an item onto a
playlist, we do an add, if you drag a playlist onto a playlist we do a
reorder.  I implemented the code for windows and OS X blindly, I'll check if
it works after this checkin.

This makes the CSS look really hairy, any ideas on how to simplify it?

Added code to remove items from playlist when they expire or are deleted.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@2670 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: eea83a206dbb54861d2d90ad4116fd413af4a9d9