Bunch of windows fixes.  Fixed delete when in an input box.  Fixed channel folder display.  Fixed Alt+F4 bug.  Made add feed dialog work with the prefill options.  Split up mp3 from mpeg associtiations in the installer.  Made the installer clean up better.  Made us not show the thumbs.db file.
Fixed a couple bugs in normalizeFeedURL.  Made validateFeedURL not work when
the URL has spaces in the middle.

Removed the javascript: feature in triggerActionURLOnUnload.  It wasn't
supported in windows or linux and it doesn't seem neccesary.  The only thing
we used it for was the external playback template and we didn't need it there.




git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@2929 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: fd26f66d054754a644c6f0917e0b775506110349