Fix for #3902.  Two weird bugs were interacting here.  The first was a bug in item.getState() that could cause it to change states when signalChange() was called, even if the change had nothing to do with the item state.  The problem was a that getState() depended on getExpiration() which depended on getState().  The only thing that prevented an infinite loop was our caching system, but you still got the weird signalChange() bug.  Fixed this by removing the call to isDownloaded() in getExpiration().  We call getSeen() anyway, which should cover it.
This caused a bug in the selection code, because setSelected() was causing the
item to be removed from the selection view (the item state was changing from
saved to unwatched).  Changed the code to add the id to currentSelection
first, then call setSelected.  If this results in the item being removed from
the view we'll do the right thing because of onRemove().



git-svn-id: https://svn.participatoryculture.org/svn/dtv/branches/Democracy-Player-0.9.0@2997 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: e4c44cb4340ee4ec5921bf94ec06a56756cc745a