Fixed bug in linux when dropping file URLs and other URLs that didn't need to be escaped in any way.
Added support for creating a new folder when adding a OPML file with multiple
channels.  Added support for dropping a file with an .opml extension.

Move filenameFromURL from the dl_daemon.downloader to download_utils.  I was
going to use this for the default folder name, but it wasn't very usefull.
Still, the move could be usefull in the future.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@3061 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: e16062d77ff747be4c155c2dd447a36659daeb2c