Implemented the changeItem speedup we discussed on the call this week.  The portable code now passes two extra arguments to changeItem: changedAttrs and newInnerHTML.  It also passes the same two extra arguments for each tuple in changeItems.  changedAttrs is a dict that stores the attribute changes for that element, if an attribute was changed the new value will be in changedAttrs, if it was removed it will have a None value in changedAttrs. newInnerHTML is the innerHTML of the element, or None if the innerHTML didn't change.  Implemented some speedups on linux, made it so things hopefully won't crash on OS X an windows.
Needed to make some template changes to make it actually work.  The problem
was that we had code like this:

<div t:repeatForView="viewWithItems">
   <div t:replaceMarkup="this.getItemXML()" />
</div>

But this doesn't get optimized, since the tag we're repeating is just a plain
<div>, the actual div for the item is inside there, so we would always change
the inner html.

Made a new option for t:repeatForView called t:repeateTemplate.  If that's set
we use that template to generate the tag we repeat.  This lets us optimize the
item DIVs with the new code.  It also lets us cleanup getItemXML() a ton,
since we don't have to have the hacks for changing the attributes all the
time.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@3581 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 788df0029d3cd7046bf97e10e5e629727babdf2d