Added resumeTime save value to item.py.  Implement saving this value in app.py (including default getCurrentTime and playFromTime functions for unimplemented platforms.)
Reimplement these functions on linux to be a bit more understandable.
Handle None in MainFrame.py.

Reworked how different renderers get embedded under linux.

Fixed a race condition in xine_impl.c where we join on a thread that is blocking on the GIL which we hold.  This was there before, but the timing changes from the resumeTime changes caused the race to be lost.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@4322 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 4b8823aeef2f9f6eb9281ed33e4f1d772c519c47