Pretty big overhaul of how we load the channel guide.  Now we don't try to download the guide at all.  Instead we load the guide-loading template which does a META refresh to the guide's URL.  This seems and cleaner to me and makes the page load quicker.  It also allows channel guide creators to have relative links on the front page and use ajax calls.
The one drawback is that we don't store the redirected URL before, so if the
channelguide homepage redirects to another page it opens in a separate
browser.  I don't think this is a problem, we just specifiy that channel
guides can't do that.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@4481 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 1858425962293557b37f5ef7341dfc30ae5973cc