Make linux argument passing to a running process work again.
We actually pass the data across as bytecode filenames decoded as latin1, and encoded as utf8.  However, on the other side we decode the utf8 and then convert to bytes using latin1.  As filenames are usually utf8, this means that files will be double utf8 encoded, but we do this in case the filenames aren't utf8.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@4609 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f608fd5eeae95438acf4f6eeed26647cd7088428