Bunch of fixes to the tab code.  We weren't saving the TabOrder object when tabs got removed.  Duh!
Added a bunch of code to deal with tab ids that are in the id list, but don't
exist in the database.  We try to clear them out when we first start up.
Also, instead of using the ID list to figure the tab 3 items higher we use the
view.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@5451 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: a74cb5ef934cabff90b4dc3da7774d45223a622a