Handle links that open in new windows
we do this by catching the "new_window" signal, making a new GtkMozEmbed object, and connecting
to the "open_uri" signal.  We then pass that URL to the openURL handler.

Note: at the moment, if a guide tries to open a URL that is part of the guide in an external
window, nothing will happen.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@5797 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: fa0ae80d5ebfdc64618713d3cedc393e312179b5