Bug #9662.  This undoes the fix in r6402 and does a different fix that doesn't break GTKX11.
There are a couple of curious things here.  I suspect the fdopen call
is wrong and that it should be dumping binary information.  I'm not
sure what the t flag does.  The initialHTML is a unicode object--that's
what it should be.  I'm not sure why we need to encode it to utf-8
in order to write it to the file.  I'd look into it more, but don't
have time right now.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@6403 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 36fb17652adab8072b8f702839f528468ad668e6