Fixes for deleting channel folders.
We had two bugs when deleting channel folders.  One is that we might also try
to delete the child items after the parent was already deleted.  The other was
that we tried to update the selection in the middle of a delete.  Both caused segfaults.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7464 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 0125a3e1f5bc10352625f2ee84f1291beccb0420