Implroved playback code.
We now handle playing back a list of videos instead of just one.  Added code to calculate the list when the user hits the play button, chooses play in the
context menu or clicks on the play icon in the item.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7582 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 2b445ead7f0d53abb76e412489559b5b573b25cb