Handle saving searches.
Added code to save the current search state.  Made the search display show the
current search engine and text when the user switches to it.

Generalized code that waited for the guide list to wait for any number of
messages.  Added CurrentSearchinfo to that list.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7787 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 62b98391aac196d0aa5dba728306abb3bd10cdc6