Some cosmetic changes but a lot of bit changes that are not cosmetic, too!
This attempts to chip away at bug #10369 where is possible for Miro to
compile against xulrunner 1.9 and then try to run against xulrunner 1.8
because the code that does the wild guessing is asstastic at best.

I also fixed some code that was borked in wild and disgusting ways, some of
which I'm pretty sure I merged in when working out xulrunner 1.9 support.
I must have been high to have checked in that crud.

I also also added a setup.py command called "test_system".  For now it
just prints out the mozilla stuff it figured out.  I'd like for it to try
importing all the things that Miro on gtkx11 uses with error messages so
that users on new systems can have a better chance of knowing whether
Miro will work or whether they're missing things.

I think I got a lot of this right, but it needs testing on systems other
than Gutsy and Hardy.  Even if it sucks on other systems, it should make
it a lot easier for packagers to package Miro.  But that's theoretical--we'll
know more as time goes on.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7885 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: dd3c798b2351a8877a9f4e74749321e6ae300a76