Fixing an issue where last video in a playlist wouldn't "stop" properly
This was related to an issue due to the position index being
incremented to a greater value than the sum length of the playlist, so
it would throw an exception once it hit update_current_resume_time.
The solution here is to simply not increment the index when we're
already at the last item in the list and stopping.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7939 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: c27cd1cca9689f5ec9cbb72c450bb0b9addec60a