Better positioning of controls.
Gave TextEntries, OptionMenus and other controls a baseline() method.  It
returns the amount of space between the bottom of the widget and the baseline
of the text it displays.  Used this to properly position controls vertically
in the preference panel.

Changed the set_size() method for OptionMenu and Button.  Now it just supports
2 sizes, small and large, instead of arbitrary sizes.

Moved OptionMenu and Button from simple.py to control.py on GTK.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8006 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: a6d7c2e12ac6e24a506e5284dbb11fade166f2f9