Fixes for the startup process fixes and dialogs.
Changed the "startup-success" and "startup-failure" signals to messages.  Made
it so we don't hange when we get a startup failure.

Changed dialogs so they aren't always transient for the main window.  This
prevents errors when we don't have a main window yet.

Refactored a bunch of Dialog classes to give them a common base class.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8245 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 7a508e55fe222aad90b744cb9857d0ca151197e4