Obscure bug which would cause validation errors
Basically, if the .url attribute in the IconCache starts with '/', it
would *always* throw a validation error on Linux, because it wouldn't
convert the filename to the required filename type on Linux (which is
type 'str') and would always return a unicode object.

A pretty rare but obscure error.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8296 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 98a1ac43e812c82ce5fbdd24d432f09e866e952c