Implemented hotspot tracking.
Implemented inital hotspot rendering for items.  Still has a bunch of issues
to work out, but it seems okay.

Fixed bug in the thumbnail positioning.  It was much more apparent when we
were switching between showing a button and not more often.

Added code on OS X to handle when a view is scrolled by it's parent Scroller
widget.  Fixed a bug that caused us not to call viewport_repositioned for
TableView objects on OS X.  Made Expander resize its child frame so it can
hold its child in before the child is placed.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8319 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 60b84bd4bcb85a24c9cd611b7774a6102808bca8