Added a display for when we can't play videos.
Refactored the way we test if we can open a file.  Instead of the
can_play_file method, the select_file method inputs 2 callbacks.  One is
called if the file is successfully opened, one is called if there is an error
opening the file.

Made some videobox widgets disable themselves while we are opening a video
(and if opening the video fails).

Made it so we don't try to save our resume time if the video was not
successfully opened.

Moved some code in OS X VideoRenderer out of teardown and into the
viewport_removed() method.  This let's us remove the widget from it's parent,
then add it to back without running into errors.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8527 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f4e206673192a73fa03c6e9408d4ab53882a96de