Implemented list view for item lists.
Created a TableColumn class that tracks the state of a single table column.
This is needed because we want to do much more complicated things for columns
in list view than we were doing before.

Added a horizontal scroller for ItemContainerWidget.  It turns on when the
user resizes the columns of the list view to overflow the window.

Canceled DnD on GTK when the drag starts from outside the content area (for
example in the column headers because the user is trying to resize them)

Added set_auto_resizes to TableView.  If set, when the table view changes
size, it's columns will also change size.  We set this for the tab list and
the normal item list, but not the list view item list.

Added a set_columns_draggable method, if set the user can reorder the columns
of a tableview by dragging them.

Moved code to strip HTML from ItemList into ItemListGroup, this way we only
need to call it once for each ItemInfo.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8562 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 9b7b006ec3f3c48d3fc2596f1d9c8dc38b31ca17