Moved video playback keyboard controls to portable.
Added support for signal responders to return True and interrupt the signal
change.  Made it so if a key-press handler returns True, we stop processing
the key on GTK and OS X.

Added the key-press signal to Window objects.  Added support for modifiers to
the key-press signal.

Removed Left and Right from the menubar.

Kept the Alt-Enter shortcut as GTK-only.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8687 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 908b2f31754bef80083efcd267e0aeb1973bd7a7