Reduce the number of InfoChanged/TabsChanged messages to the frontend.
The optimization is simple: check if the info object actually changed.  There
are lots of reasons that signalChange() gets called, but only some of them
actually affect the ItemInfo/ChannelInfo object that we are sending to the
frontend.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8795 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 6509f412cd7055ccbd60306d14ad8df9f9b1ad0b