Style changes for selected items (#10938)
Refactored the draw_background_details method to make it use the usual codepath
as the normal background parts.

Made it so the selected border is inside where the normal border would be,
rather than surrounding it.  This makes the buttons for the flap look a less
crammed.


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8801 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: d2badebdfff368dc41c26986c0b115e0ab13095d