reenable themes on Windows, and fix a platform-agnostic bug
Themes were broken on Windows because nothing was checking for the --theme
commandline option.  I've added the code which checks for that and sets the
theme name.

Also, a method name changed in the OPML importer, which caused an exception
when loading feeds for a theme.

Refs #10852

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8855 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 74fc7599cff9af844c71fdfaf6e2858e36bb08e7