Better way to handle items getting removed from the new videos list.
Instead of the method used in r8660, I added a new message called
TrackItemsManually.  It allows the frontend to specify a list of ids to track.
The playlist uses this message to track it's items, regardless of if they
exist in the view that we're playing from or not.

This is better because we keep tracking the items even after they're out of
the feed.  For example, when playing items from the new videos tab, we now see
that they get marked unwatched, so the user has a chance to keep them from the
video details widget (#11063).

I think it also makes more sense in a bunch of edge cases.  For example,
playing a video, then deleting the feed that it came from, but keeping the
items.  Before we would stop playing, but now we keep going.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8864 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: a719188767a5663c9def76fdaa22db262bc07d94