Fixes for hetrogenous item list context menus.
Before the actions would happen for every item in the list, rather than the
portion of the list that the action was for.  For example, if you had a list
with 1 downloading item and 1 paused item, if you clicked "Cancel Download"
under the "1 Downloading Item" it would apply to both the downloading and
paused items.

Added an option to cancel paused downloads.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8888 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 4dc6f9fd371b295a0914918958b6d636f59231a2