allow themes to limit/reorder/include new search engines
This commit introduces a new preference, SEARCH_ORDERING, which can be used in a theme to adjust the search engines (including 'All') that are shown to the user.  It's a comma-separated list of search engine names, with an '*' as an optional last item which includes the rest of the engines.

This also adds code which looks for search engines in the theme directory, so that themes can include their own search engines.  Icons for those themes are looked up in the theme directory as well.  The code for looking up icons for theme searches has only been implemented for GTK, but this commit shouldn't break current code on OS X.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8897 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 54d460984cd1edf17a97c3125aedfbc00a10d148