Better fix for #10751.
When we detect that a two torrent downloads have the same info hash, we update
one of the items so that it shares a downloader with another.

This means we have to change how we re-link items to downloade, since using
the URL won't work.  Instead, we give items a downloader_id attribute and use
that.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8963 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f2d3950ef2826a399e2ecb6a80fa07de9c1664b2