Bug #11148: code to handle loading icon for themed Guides
For the Miro Guide, we want to use the pretty icon that we ship with Miro and
always use the name "Miro Guide" (suitably translated).  For themes, we want to
let the guide decide its own icon and name.  This checkin defers to the Guide
unless it's the Miro Guide, in which case we use the defaults.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8991 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: d7fced01542bd744d4a405ff2447dde2272d0563