Fix for selecting static tabs.
Before when selecting a new static tab, we would unselect the current tab ev
when it was the tab to be selected.  This would mess things up when you
started a search when the search tab was already selected.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9065 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 2c376ad7ec1f47b7392f05e2f94d0295b1d25a82