This re-works frontends to provide a common interface for them all on Linux. This interface could/should be extended to other platforms, but that's a project for a different day.
A "frontend" consists of an application module in a package that has
a run_application function in it.  This function sets up the frontend.
See cli/application.py and shell/application.py for examples.

This allows frontends to be anywhere in the PYTHONPATH--they don't have
to be in the miro core codebase.

To run a different frontend, use the --frontend switch with the value
being the package the application module is in.

Examples:

   --frontend=miro.plat.frontends.widgets
   --frontend=miro.frontends.cli

Additionally, to save some typing, Miro will try miro.frontends.FRONTEND and
miro.plat.frontends.FRONTEND automatically allowing for the two examples to
be:

   --frontend=widgets
   --frontend=cli

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9206 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: bf9029f3580f1a86f51793e4db1522e286845a1b