Made the menu code more sane.
Now when the selection changes we always call app.menu_manager.update_menus().
This simpifies things on both sides of the equation.  For MenuManager, it's
simpler to handle everything in one place, rather than worry about the order
that methods get called in.  For other code, it's simpler to call
update_menus() and know that MenuManager will do the Right Thing.

This change also fixes it so that the play button and the play/pause menu item
are always in sync.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9295 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 3af7ec3007ded029a6e0759477c5c1ca977b0308