Try to load objects using ids before using the entire row.
This is an optimization for the sql views.  It's much quicker to load only the
row ids from sqlite rather than the entire rows (I think the documentation
says it's 1 binary search instead of 2).  If the objects are already loaded in
memory, we only need to select their ids, so the query is much quicker.  If
the objects are not already loaded in memory, than the query will probably be
slower, but I think this is a much smaller difference because loading a bunch
of rows from a set of row ids should be fast.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9451 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: cb97b973ed72621372b920b61789257e5ab940f9