Fix for #11756
The problem was for that RSSMultiFeedImpl kept a list of the old items from
each subfeed when they were updating.  If 2 subfeeds contained the same item
and that item gets truncated, then we would try to remove it twice.  Changed
the list to a set and added a unit test for this case.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9569 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f2e8cf795483efbd5397d7673d4861f27296b095