Hack to handle packing lots of buttons for torrent folders.
It's possible that we will pack the "display contents", "keep" and "remove"
buttons next to each other for torrent folders.  There also could be an
expiring/unwatched emblem next to the play button.  This takes up a lot of
horizontal space.  If that happens, and the user has made the Miro window not
that wide, then we cover up the emblem with the buttons.

I think this also makes the emblem packing code a bit simpler, although it's
still a semi-mess.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9664 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 54d6a9d1316ece007e56ddde5d192ee3007acf57